feat: change e2e test to autobinding
