+++ .gitmodules
@@ -5,14 +5,14 @@
 	path = libs/providers/flagd-web/schemas
 	url = https://github.com/open-feature/schemas
 [submodule "libs/providers/flagd/spec"]
-	path = libs/providers/flagd/spec
-	url = https://github.com/open-feature/spec.git
-[submodule "libs/providers/flagd-web/spec"]
-	path = libs/providers/flagd-web/spec
-	url = https://github.com/open-feature/spec.git
+  path = libs/providers/flagd/spec
+  url = https://github.com/open-feature/spec.git
 [submodule "libs/shared/flagd-core/flagd-schemas"]
 	path = libs/shared/flagd-core/flagd-schemas
 	url = https://github.com/open-feature/flagd-schemas.git
-[submodule "libs/providers/flagd/flagd-testbed"]
-	path = libs/providers/flagd/flagd-testbed
-	url = https://github.com/open-feature/flagd-testbed.git
+[submodule "libs/shared/flagd-core/test-harness"]
+	path = libs/shared/flagd-core/test-harness
+	url = https://github.com/open-feature/flagd-testbed
+[submodule "libs/shared/flagd-core/spec"]
+	path = libs/shared/flagd-core/spec
+	url = https://github.com/open-feature/spec
