--- libs/providers/flagd-web/src/e2e/constants.ts
@@ -1,3 +1,0 @@
-export const FLAGD_NAME = 'flagd-web';
-
-export const IMAGE_VERSION = 'v0.5.6';
