--- libs/providers/flagd-web/src/e2e/features/.gitignore
@@ -1,1 +1,0 @@
-*.feature
