+++ libs/providers/flagd-web/tsconfig.spec.json
@@ -6,5 +6,5 @@
     "types": ["jest"],
     "allowJs": true,
   },
-  "include": ["jest.config.ts", "**/*.test.ts", "**/*.spec.ts", "**/*.d.ts"]
+  "include": ["jest.config.ts", "**/*.test.ts", "**/*.spec.ts", "**/*.d.ts","src/e2e/"]
 }
