--- libs/providers/flagd/src/e2e/features/.gitignore
@@ -1,1 +1,0 @@
-*.feature
