+++ libs/providers/flagd/src/e2e/tests/rpc-reconnect.spec.ts
@@ -1,31 +1,22 @@
 import assert from 'assert';
 import { OpenFeature } from '@openfeature/server-sdk';
 import { FlagdProvider } from '../../lib/flagd-provider';
+import { GenericContainer, StartedTestContainer } from 'testcontainers';
+import { autoBindSteps, loadFeature } from 'jest-cucumber';
 import {
-  E2E_CLIENT_NAME,
   FLAGD_NAME,
-  UNSTABLE_CLIENT_NAME,
-  UNAVAILABLE_CLIENT_NAME,
   IMAGE_VERSION,
-} from '../constants';
-import { evaluation } from '../step-definitions/evaluation';
-import { GenericContainer, StartedTestContainer, TestContainer } from 'testcontainers';
-import { flagd } from '../step-definitions/flagd';
-import { flagdJsonEvaluator } from '../step-definitions/flagd-json-evaluator';
-import { flagdRecconnectUnstable } from '../step-definitions/flagd-reconnect.unstable';
+  UNAVAILABLE_CLIENT_NAME,
+  UNSTABLE_CLIENT_NAME,
+  reconnectStepDefinitions,
+  GHERKIN_FLAGD_RECONNECT_FEATURE,
+} from '@openfeature/flagd-core';
 
 // register the flagd provider before the tests.
 async function setup() {
   const containers: StartedTestContainer[] = [];
 
   console.log('Setting flagd provider...');
-
-  const stable = await new GenericContainer(`ghcr.io/open-feature/flagd-testbed:${IMAGE_VERSION}`)
-    .withExposedPorts(8013)
-    .start();
-  containers.push(stable);
-  OpenFeature.setProvider(E2E_CLIENT_NAME, new FlagdProvider({ cache: 'disabled', port: stable.getFirstMappedPort() }));
-
   const unstable = await new GenericContainer(`ghcr.io/open-feature/flagd-testbed-unstable:${IMAGE_VERSION}`)
     .withExposedPorts(8013)
     .start();
@@ -35,14 +26,6 @@ async function setup() {
     new FlagdProvider({ cache: 'disabled', port: unstable.getFirstMappedPort() }),
   );
   OpenFeature.setProvider(UNAVAILABLE_CLIENT_NAME, new FlagdProvider({ cache: 'disabled', port: 8015 }));
-  assert(
-    OpenFeature.getProviderMetadata(E2E_CLIENT_NAME).name === FLAGD_NAME,
-    new Error(
-      `Expected ${FLAGD_NAME} provider to be configured, instead got: ${
-        OpenFeature.getProviderMetadata(E2E_CLIENT_NAME).name
-      }`,
-    ),
-  );
   assert(
     OpenFeature.getProviderMetadata(UNSTABLE_CLIENT_NAME).name === FLAGD_NAME,
     new Error(
@@ -59,11 +42,13 @@ async function setup() {
       }`,
     ),
   );
+
   console.log('flagd provider configured!');
   return containers;
 }
 
-describe('rpc', () => {
+jest.setTimeout(30000);
+describe('in process', () => {
   let containers: StartedTestContainer[] = [];
   beforeAll(async () => {
     containers = await setup();
@@ -78,8 +63,7 @@ describe('rpc', () => {
       }
     }
   });
-  evaluation();
-  flagd();
-  flagdJsonEvaluator();
-  flagdRecconnectUnstable();
+
+  const features = [loadFeature(GHERKIN_FLAGD_RECONNECT_FEATURE)];
+  autoBindSteps(features, [reconnectStepDefinitions]);
 });
