+++ libs/providers/flagd/tsconfig.lib.json
@@ -6,5 +6,5 @@
     "types": []
   },
   "include": ["**/*.ts"],
-  "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
+  "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts", "./src/e2e"]
 }
