+++ libs/shared/flagd-core/src/e2e/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './step-definitions';
