+++ libs/shared/flagd-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/flagd-core';
 export * from './lib/feature-flag';
 export * from './lib/storage';
+export * from './e2e';
